backend/requirements.txt
Flask==2.3.2
boto3
python-dotenv
flask_cors
mangum
werkzeug